refactor(workspace): split markdown materializer into formatting + factory
markdown.ts was a god file mixing pure formatting (toMarkdown, markdown())
with the 200-line stateful filesystem materializer. Split into:
- markdown.ts: pure formatting helpers + SerializeResult type
- materializer.ts: createMarkdownMaterializer factory + I/O

Also deletes materializer/types.ts (MaterializerExtension had zero
consumers) and updates the spec to reflect that createMaterializer
rename was rejected in favor of keeping createMarkdownMaterializer.
